- made service functions static, removed userSnakeService function into userService, changed routes folder to controllers, changed statsService folder to statsSingleton, added get functions to userService and statsSingleton, statsController no longer uses any models
